FIX: remove subscriptions on force close
