+++ CHANGELOG.md
@@ -1,0 +1,32 @@
+# Next Version
+
+### ✅ Features & Improvements
+
+- ...
+
+### 🐛 Bug fixes
+
+- ...
+
+# 2.0.0
+
+### 💣 Breaking Changes
+
+- Changed class name from `neurosity_sdk` to `NeurositySDK`
+- Moved `example.py` from `./neurosity/` to `./examples/`
+- Dropped `pyrebase` in favor of `firebase-rest-api`
+
+### ✅ Features & Improvements
+
+- Added this changelog!
+- Added `.devcontainer` to standardize development
+- Added `requirements.txt` and `dev-requirements.txt`
+- Added section on developing this project to the README
+
+### 🐛 Bug fixes
+
+- Automatically remove subscriptions on forced terminal kills, keyboard kills, and other system faults
+
+# 1.0.2
+
+Initial Release
