+++ neurosity/__init__.py
@@ -1,2 +1,2 @@
-from neurosity.config import PyRebase
-from neurosity.neurosity import neurosity_sdk
+from .config import FirebaseConfig
+from .neurosity import *
