+++ neurosity/config.py
@@ -1,14 +1,16 @@
-class PyRebase:
+class FirebaseConfig:
     PRODUCTION = {
       "apiKey": "AIzaSyB0TkZ83Fj0CIzn8AAmE-Osc92s3ER8hy8",
       "authDomain": "neurosity-device.firebaseapp.com",
       "databaseURL": "https://neurosity-device.firebaseio.com",
-      "storageBucket": "neurosity-device.appspot.com"
+      "storageBucket": "neurosity-device.appspot.com",
+      "projectId": "neurosity-device"
     }
 
     STAGING = {
       "apiKey": "AIzaSyDfw8CFZBrcWyqS23888ULvoKru7fnlz5Q",
       "authDomain": "neurosity-device-staging.firebaseapp.com",
       "databaseURL": "https://neurosity-device-staging-default-rtdb.firebaseio.com",
-      "storageBucket": "neurosity-device-staging.appspot.com"
+      "storageBucket": "neurosity-device-staging.appspot.com",
+      "projectId": "neurosity-device"
     }
