+++ neurosity/neurosity.py
@@ -1,29 +1,32 @@
+import atexit
 import firebase
 import signal
 
-from neurosity.config import PyRebase
+from .config import FirebaseConfig
 
-class neurosity_sdk:
+class NeurositySDK:
     def __init__(self, options):
         if ("device_id" not in options):
             raise ValueError("Neurosity SDK: A device ID is required to use the SDK")
 
         options.setdefault("environment", "production")
         self.options = options
-        pyrebase_config = PyRebase.STAGING if options["environment"] == "staging" else PyRebase.PRODUCTION
+        pyrebase_config = FirebaseConfig.STAGING if options["environment"] == "staging" else FirebaseConfig.PRODUCTION
         self.firebase = firebase.initialize_app(pyrebase_config)
         self.auth = self.firebase.auth()
         self.db = self.firebase.database()
         self.subscription_ids = []
 
-        # register a signal handler for SystemExit and KeyboardInterrupt
-        signal.signal(signal.SIGTERM, self.exit_handler)
-        signal.signal(signal.SIGINT, self.exit_handler)
+        # For a normal exit
+        atexit.register(self.exit_handler)
 
-    def exit_handler(self):
+        # register a signal handler for Forced Terminal Kills
+        signal.signal(signal.SIGHUP, self.exit_handler)
+
+    def exit_handler(self, signum=None, frame=None):
         self.remove_client()
         self.remove_all_subscriptions()
-
+        
     def get_server_timestamp(self):
         return {".sv": "timestamp"}
